final commit for the project

+++ app/Http/Controllers/API/UserController.php
@@ -27,7 +27,12 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        // Querying Data from Database for Admin and Author Only.
+        // $this->authorize('isAdmin');
+        if(\Gate::allows('isAdmin') || \Gate::allows('isAuthor')){
+            return User::latest()->paginate(5); 
+        }
+        
     }
 
     /**
@@ -135,10 +140,26 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
 
         // delete User
           $user->delete();
         return ['message' => 'User Deleted'];
+    }
+
+    public function search()
+    {
+
+        if ($search = \Request::get('q')) {
+            $users = User::where(function ($query) use ($search) {
+                $query->where('name', 'LIKE', "%$search%")
+                    ->orWhere('email', 'LIKE', "%$search%");
+            })->paginate(20);
+        } else {
+            $users = User::latest()->paginate(5);
+        }
+
+        return $users;
     }
 }
